color working as type
